Split timeline service.
